moved to f32

+++ src/camera.rs
@@ -6,7 +6,7 @@ pub struct Camera {
     lower_left_corner: Point3,
     horizontal: Vector3,
     vertical: Vector3,
-    lens_radius: f64,
+    lens_radius: f32,
     u: Vector3,
     v: Vector3,
 }
@@ -14,12 +14,12 @@ pub struct Camera {
 
 impl Camera {
 
-    pub fn new(look_from: Point3, look_at: Point3, vup: Vector3, vfov: f64, aspect_ratio: f64, aperture: f64, focus_dist: f64) -> Camera {
+    pub fn new(look_from: Point3, look_at: Point3, vup: Vector3, vfov: f32, aspect_ratio: f32, aperture: f32, focus_dist: f32) -> Camera {
         let theta = vfov.to_radians();
         let h = (theta / 2.0).tan();
 
-        let viewport_height: f64 = 2.0 * h;
-        let viewport_width: f64 = aspect_ratio * viewport_height;
+        let viewport_height: f32 = 2.0 * h;
+        let viewport_width: f32 = aspect_ratio * viewport_height;
 
         let w = (look_from - look_at).normalized();
         let u = Vector3::cross(&vup, &w).normalized();
@@ -37,7 +37,7 @@ impl Camera {
 
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd: Vector3 = self.lens_radius *Vector3::random_in_unit_disk();
         let offset: Vector3 = self.u * rd.x + self.v * rd.y;
