+++ src/rays.rs
@@ -14,7 +14,7 @@ impl Ray {
         Ray { origin, direction }
     }
 
-    pub fn at(self, t: f64) -> Point3 {
+    pub fn at(self, t: f32) -> Point3 {
         self.origin + t * self.direction
     }
